Move libs to libs/celestia-app
