+++ libs/celestia-app/namespace/consts.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"math"
 
-	"github.com/rollkit/rollkit/libs/appconsts"
+	"github.com/rollkit/rollkit/libs/celestia-app/appconsts"
 )
 
 const (
