+++ libs/celestia-app/namespace/namespace_test.go
