+++ libs/celestia-app/namespace/random_blob.go
