+++ libs/celestia-app/shares/doc.go
