+++ libs/celestia-app/shares/info_byte.go
@@ -3,7 +3,7 @@ package shares
 import (
 	"fmt"
 
-	"github.com/rollkit/rollkit/libs/appconsts"
+	"github.com/rollkit/rollkit/libs/celestia-app/appconsts"
 )
 
 // InfoByte is a byte with the following structure: the first 7 bits are
